Script to assign db containers to freezer addresses
Create a script to assign db containers to freezer addresses. This
script creates the necessary freezer addresses if they do not yet exist.
This script should be used as the final step in migrating a SDB v1 to
SDB v2. After running this script all containers in the db will be
assigned to the proper container.