Generate a SDB backup each time the app is run
This does not necessarily create a back up after each move or upload,
bc these funs can be used outside of the app.
Backups are also generated twice a day, midnight and noon.
UploadCSVs and Upload Moves should be stored.
And other measures may need to be taken in order to ensure backing up
the DB is robust.