Patched VisionMate empty row uploadCSV
Sisi tried to upload a visionmate upload csv file but it did not work,
presumably bc empty tubes are noted literally as the string "No Tube", rather than "".
This commit servers as a patch to this issue. I may have to check how empty Traxcer rows are handled during uploads